feat: add balance module
A balance module provides access to external balance-handling functionality, which is contained within the Substrate layer in our case.

If a script wants to execute a transfer, a transaction which contains the script must receive two parameters:
- signer - A proof that the account has signed the transaction.
- cheque_amount - How many funds is the signer account allowing a script to transfer from that account.

The provided amount remains in the signer account if the script doesn't perform any transfer.